[trunk] Support double precision records for display
Double precision samples were converted to float for display purposes
and could be shown in many cases. But if a double precision value does
not fit into a float it would lead to incorrect plots. This commit
allows to render double precision records with full precision.

Some Qt signals and slots were required to change its signature from
float to double which is not checked at compile time. This would have
an effect on external code which uses the affected signals/slots. See
version.h for more information.

+++ libs/seiscomp/gui/core/recordwidget.h
@@ -232,23 +232,23 @@ class SC_GUI_API RecordWidget : public QWidget {
 			Box
 		};
 
-		typedef Math::Filtering::InPlaceFilter<float> Filter;
+		typedef Math::Filtering::InPlaceFilter<double> Filter;
 
 		struct Trace {
 			Trace() = default;
 
-			float                     dyMin{0};   // Data minimum value
-			float                     dyMax{0};   // Data maximum value
-			float                     dOffset{0}; // Data offset
-			float                     absMax{0};
+			double                    dyMin{0};   // Data minimum value
+			double                    dyMax{0};   // Data maximum value
+			double                    dOffset{0}; // Data offset
+			double                    absMax{0};
 			int                       pyMin{0};   //
 			int                       pyMax{0};
-			float                     fyMin{-1};
-			float                     fyMax{1};
+			double                    fyMin{-1};
+			double                    fyMax{1};
 
-			float                     yMin{0};
-			float                     yMax{0};
-			float                     yOffset{0}; // The used offset
+			double                    yMin{0};
+			double                    yMax{0};
+			double                    yOffset{0}; // The used offset
 
 			float                     timingQuality{-1};
 			int                       timingQualityCount{0};
@@ -380,7 +380,7 @@ class SC_GUI_API RecordWidget : public QWidget {
 		float timingQuality(int slot) const;
 
 		double timeScale() const { return _pixelPerSecond; }
-		float amplScale() const { return _amplScale; }
+		double amplScale() const { return _amplScale; }
 		double tmin() const { return _tmin; }
 		double tmax() const { return _tmax; }
 		double amplMin() const { return _amplitudeRange[0]; }
@@ -400,11 +400,11 @@ class SC_GUI_API RecordWidget : public QWidget {
 
 		//! Returns the amplitude range of the available data in the normalization
 		//! window
-		QPair<float,float> amplitudeDataRange(int slot) const;
+		QPair<double,double> amplitudeDataRange(int slot) const;
 
 		//! Returns the amplitude range of the trace slot where first corresponds
 		//! to the lower pixel row and second to the upper pixel row
-		QPair<float,float> amplitudeRange(int slot) const;
+		QPair<double,double> amplitudeRange(int slot) const;
 
 		void ensureVisibility(const Seiscomp::Core::Time &time, int pixelMargin);
 	
@@ -519,9 +519,6 @@ class SC_GUI_API RecordWidget : public QWidget {
 		//! Maps a widget position to a time
 		Core::Time unmapTime(int x) const;
 
-		//bool setGain(float);
-		//float gain() const;
-
 		const DataModel::WaveformStreamID& streamID() const;
 
 		void setSlotCount(int);
@@ -547,7 +544,7 @@ class SC_GUI_API RecordWidget : public QWidget {
 		void setSelected(double t1, double t2);
 		void setSelected(Seiscomp::Core::Time t1, Seiscomp::Core::Time t2);
 
-		void setScale(double, float=0);
+		void setScale(double, double=0);
 		void setTimeScale(double);
 
 		void setTimeRange(double, double);
@@ -562,7 +559,7 @@ class SC_GUI_API RecordWidget : public QWidget {
 		void setAlignment(Seiscomp::Core::Time t);
 		void alignOnMarker(const QString& text);
 
-		void setAmplScale(float);
+		void setAmplScale(double);
 	
 		void enableFiltering(bool enable);
 		void setGridSpacing(double, double, double);
@@ -666,7 +663,7 @@ class SC_GUI_API RecordWidget : public QWidget {
 
 		virtual void createPolyline(int slot, AbstractRecordPolylinePtr &polyline,
 		                            RecordSequence const *, double pixelPerSecond,
-		                            float amplMin, float amplMax, float amplOffset,
+		                            double amplMin, double amplMax, double amplOffset,
 		                            int height, bool optimization, bool highPrecision);
 		virtual const double *value(int slot, const Seiscomp::Core::Time&) const;
 
@@ -754,7 +751,7 @@ class SC_GUI_API RecordWidget : public QWidget {
 		double               _tmax;            // time range max
 		double               _smin, _smax;     // selection
 		double               _pixelPerSecond;
-		float                _amplScale;       // pixel per amplitude unit (0=normalize)
+		double               _amplScale;       // pixel per amplitude unit (0=normalize)
 		double               _gridHSpacing[2];
 		double               _gridHOffset;
 
@@ -763,7 +760,7 @@ class SC_GUI_API RecordWidget : public QWidget {
 		double               _gridVOffset;
 		double               _gridVScale;
 
-		float                _amplitudeRange[2];
+		double               _amplitudeRange[2];
 		bool                 _useFixedAmplitudeRange;
 		bool                 _useMinAmplitudeRange;
